+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+### 0.1.0-beta
+
+- New command to generate a Lightning Web Component
+- New command to erase stored credentials on demand
+
 ### 0.0.1-beta
 
 First beta release supporting Apex code generation, modification, and tests.